requirements.txt
Django==2.2.5
django-autoslug==1.9.6
django-ckeditor==5.7.1
django-jet==1.0.8
django-js-asset==1.2.2
Pillow==6.2.0
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
